refactor: give token/ and scene/ foundry/ + logic/ folder parity
Mirror the actor/item/combatant layout for the two remaining document types
that have a logic class.

- token/: SohlTokenDocument → token/foundry/, SohlTokenDocumentLogic →
  token/logic/.
- scene/: SohlScene + SohlSceneDataModel + SohlSceneConfig → scene/foundry/,
  SohlSceneLogic → scene/logic/.
- Add token/logic/** and scene/logic/** to the ESLint Foundry-free zone and the
  purity smoke test (both logic modules are pure today; now enforced).
- Update all importers to the new paths.

No behavior change. (Effect remains a single foundry-layer file with no logic
class; chat is glue helpers, not a document type — both left as-is.)
