+++ src/core/FoundryHelpers.ts
@@ -14,13 +14,13 @@
 import { SimpleRoll } from "@src/utils/SimpleRoll";
 import { FilePath, toSanitizedHTML, HTMLString } from "@src/utils/helpers";
 import type { SohlItem } from "@src/document/item/foundry/SohlItem";
-import type { SohlTokenDocument } from "@src/document/token/SohlTokenDocument";
-import type { SohlScene } from "@src/document/scene/SohlScene";
+import type { SohlTokenDocument } from "@src/document/token/foundry/SohlTokenDocument";
+import type { SohlScene } from "@src/document/scene/foundry/SohlScene";
 import type { SohlLogic } from "@src/core/SohlLogic";
 import type { SohlCombatant } from "@src/document/combatant/foundry/SohlCombatant";
 import type { CombatantLogic } from "@src/document/combatant/logic/CombatantLogic";
 import type { SohlActor } from "@src/document/actor/foundry/SohlActor";
-import type { SohlTokenDocumentLogic } from "@src/document/token/SohlTokenDocumentLogic";
+import type { SohlTokenDocumentLogic } from "@src/document/token/logic/SohlTokenDocumentLogic";
 
 /**
  * Foundry VTT runtime shim.
