+++ src/core/SohlSpeaker.ts
@@ -19,7 +19,7 @@ import {
 } from "@src/utils/constants";
 import { registerKind } from "@src/utils/kindRegistry";
 import type { SohlActor } from "@src/document/actor/foundry/SohlActor";
-import type { SohlTokenDocument } from "@src/document/token/SohlTokenDocument";
+import type { SohlTokenDocument } from "@src/document/token/foundry/SohlTokenDocument";
 import { FilePath, isFilePath, HTMLString } from "@src/utils/helpers";
 import { SimpleRoll } from "@src/utils/SimpleRoll";
 import { SohlSpeakerStyle } from "@src/utils/constants";
