+++ src/document/scene/foundry/SohlSceneConfig.ts
