+++ src/domain/result/SuccessTestResult.ts
@@ -13,7 +13,7 @@
 
 import { MasteryLevelModifier } from "@src/domain/modifier/MasteryLevelModifier";
 import { registerKind } from "@src/utils/kindRegistry";
-import type { SohlTokenDocument } from "@src/document/token/SohlTokenDocument";
+import type { SohlTokenDocument } from "@src/document/token/foundry/SohlTokenDocument";
 import type { SohlContextMenu } from "@src/utils/SohlContextMenu";
 import type { SohlItem } from "@src/document/item/foundry/SohlItem";
 import { SohlSpeaker } from "@src/core/SohlSpeaker";
